Changes after code review

+++ if.py
@@ -1,12 +1,12 @@
 user_input = input("Please enter your age ")
 try:
     reply = ""
-    user_age = float(user_input)
-    if user_age <=6:
+    user_age = int(user_input)
+    if user_age <= 6:
         reply = "Kindergarten"
-    elif user_age <=17:
+    elif user_age <= 17:
         reply = "School"
-    elif user_age <=23:
+    elif user_age <= 23:
         reply = "Univercity"
     else:
         reply = "Job"
